выносит рендер в отдельный модуль
