+++ src/utils/utils.js
@@ -1,9 +1,4 @@
-import {constants} from "./constants.js";
-
-const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
+import {constants} from "../constants.js";
 
 const getRandomIntegerNumber = (min, max) => {
   return Math.floor(min + Math.random() * (max + 1 - min));
@@ -70,24 +65,6 @@ const generateRandomDateComment = () => {
   return `${result.getFullYear()}/${generateCustomData(result.getMonth())}/${generateCustomData(result.getHours())}:${generateCustomData(result.getMinutes())}`;
 };
 
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-  return newElement.firstChild;
-};
-
-const render = (container, element, place = RenderPosition.BEFOREEND) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-
 const utils = {
   getRandomBoolean,
   getRandomIntegerNumber,
@@ -97,9 +74,6 @@ const utils = {
   generateRandomDateComment,
   castTimeFormat,
   getRandomMovieLength,
-  createElement,
-  RenderPosition,
-  render,
 };
 
 export {utils};
